+++ promptukit/exams/__init__.py
@@ -3,6 +3,7 @@
 
 __all__ = [
     "create_exam",
+    "create_exam_md",
     "create_pub_quiz",
     "create_pptx",
 ]